backdjango/requirements.txt
django~=4.1.4
pillow
psycopg2
djangorestframework
markdown
django-filter
